express_cut/requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.1
gunicorn==20.0.4
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
waitress==1.4.4
drf-yasg==1.17.1
django-cors-headers==3.5.0
djangorestframework-jwt==1.11.0
